requirements.txt
asgiref==3.4.1
Django==4.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-mptt==0.13.4
gunicorn==20.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
